requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# System monitoring
psutil==5.9.6

# Redis
redis==5.0.1

# Vector store & embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# LLM APIs
google-generativeai==0.3.2
huggingface-hub==0.20.3
transformers==4.40.2
torch==2.1.0

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.2
requests==2.31.0

# Day 5: Real-time features
websockets==12.0
circuitbreaker==1.4.0

# AWS SDK (for future use, compatible version)
boto3==1.34.0
websockets>=12.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
psycopg2-binary==2.9.9
